bias_lines {NeuralNetTools}	R Documentation

Plot connection weights for bias lines

Description

Plot connection weights for bias lines in plotnet

Usage

bias_lines(bias_x, bias_y, mod_in, nid, rel_rsc, all_out, pos_col, neg_col,
  struct, y_names, x_range, y_range, layer_x, line_stag, max_sp)

Arguments

bias_x
numeric vector x axis locations for bias lines

bias_y
numeric vector y axis locations for bias lines

mod_in
neural network model object

nid
logical value indicating if neural interpretation diagram is plotted, default TRUE

rel_rsc
numeric indicating the scaling range for the width of connection weights in a neural interpretation diagram. Default is NULL for no rescaling.

all_out
chr string indicating names of response variables for which connections are plotted, default all

pos_col
chr string indicating color of positive connection weights, default 'black'

neg_col
chr string indicating color of negative connection weights, default 'grey'

struct
numeric vector for network structure

y_names
chr string for names of output variables

x_range
numeric of x axis range for base plot

y_range
numeric of x axis range for base plot

layer_x
numeric indicating locations of layers on x axis

line_stag
numeric value that specifies distance of connection weights from nodes

max_sp
logical indicating if space is maximized in plot

[Package NeuralNetTools version 1.5.2 Index]