lekgrps {NeuralNetTools}	R Documentation

Create optional barplot for lekprofile groups

Description

Create optional barplot of constant values of each variable for each group used with lekprofile

Usage

lekgrps(grps, position = "dodge", grp_nms = NULL)

Arguments

grps
data.frame of values for each variable in each group used to create groups in lekprofile

position
chr string indicating bar position (e.g., 'dodge', 'fill', 'stack'), passed to geom_bar

grp_nms
optional chr string of alternative names for groups in legend

Value

A ggplot object

Examples

## enters used with kmeans clustering
x <- neuraldat[, c('X1', 'X2', 'X3')]
grps <- kmeans(x, 6)$center

lekgrps(grps)

[Package NeuralNetTools version 1.5.2 Index]