neuraldat {NeuralNetTools}	R Documentation

Simulated dataset for function examples

Description

A simulated dataset of 2000 observations containing two response variables and three explanatory variables. Explanatory variables were sampled from a standard normal distribution. Response variables were linear combinations of the explanatory variables. The response variables Y1 and Y2 are standardized from 0 to 1.

Usage

neuraldat

Format

A data frame with 2000 rows and 5 variables:

Y1

numeric

Y2

numeric

X1

numeric

X2

numeric

X3

numeric

...

[Package NeuralNetTools version 1.5.2 Index]