Plot bias points

Plot bias points in plotnet

bias_points(
  bias_x,
  bias_y,
  layer_name,
  node_labs,
  x_range,
  y_range,
  circle_cex,
  cex_val,
  bord_col,
  circle_col
)

Arguments

bias_x

numeric vector of values for x locations

bias_y

numeric vector for y location

layer_name

string indicating text to put in node

node_labs

logical indicating of node labels are included

x_range

numeric of x axis range for base plot

y_range

numeric of y axis range for base plot

circle_cex

numeric value indicating size of nodes, default 5

cex_val

numeric value indicating size of text labels, default 1

bord_col

chr string indicating border color around nodes, default 'lightblue'

circle_col

chr string indicating color of nodes