Plot neural network nodes

Plot neural network nodes in plotnet

layer_points(
  layer,
  x_loc,
  x_range,
  layer_name,
  cex_val,
  circle_cex,
  bord_col,
  in_col,
  node_labs,
  line_stag,
  var_labs,
  x_names,
  y_names,
  ...
)

Arguments

layer

specifies which layer, integer from struct

x_loc

indicates x location for layer, integer from layer_x

x_range

numeric for total range of x-axis

layer_name

string indicating text to put in node

cex_val

numeric indicating size of point text

circle_cex

numeric indicating size of circles

bord_col

chr string indicating border color around nodes, default lightblue

in_col

chr string indicating interior color of nodes

node_labs

logical indicating if node labels are to be plotted

line_stag

numeric indicating distance between of text from nodes

var_labs

chr string for variable labels

x_names

chr string for alternative names of input nodes

y_names

chr string for alternative names of output nodes

...

values passed to get_ys