Predicted values for Lek profile method

Get predicted values for Lek Profile method, used iteratively in lekprofile

pred_sens(mat_in, mod_in, var_sel, step_val, grps, ysel)

Arguments

mat_in
data.frame of only the explanatory variables used to create model

mod_in
any model object with a predict method

var_sel
chr string of explanatory variable to select

step_val
number of values to sequence range of selected explanatory variable

grps
matrix of values for holding explanatory values constant, one column per variable and one row per group

ysel
chr string of response variable names for correct labelling

Value

A list of predictions where each element is a data.frame with the predicted value of the response and the values of the explanatory variable defined by var_sel. Each element of the list corresponds to a group defined by the rows in grps at which the other explanatory variables were held constant.

Details

Gets predicted output for a model's response variable based on matrix of explanatory variables that are restricted following Lek's profile method. The selected explanatory variable is sequenced across a range of values. All other explanatory variables are held constant at the values in grps.

See also

Examples

## using nnet

library(nnet)

data(neuraldat) 
set.seed(123)

mod <- nnet(Y1 ~ X1 + X2 + X3, data = neuraldat, size = 5)

#> # weights:  26
#> initial  value 259.012592 
#> iter  10 value 0.986480
#> iter  20 value 0.225311
#> iter  30 value 0.139585
#> iter  40 value 0.098961
#> iter  50 value 0.038200
#> iter  60 value 0.022839
#> iter  70 value 0.013774
#> iter  80 value 0.008530
#> iter  90 value 0.005172
#> iter 100 value 0.003044
#> final  value 0.003044 
#> stopped after 100 iterations

mat_in <- neuraldat[, c('X1', 'X2', 'X3')]
grps <- apply(mat_in, 2, quantile, seq(0, 1, by = 0.2))

pred_sens(mat_in, mod, 'X1', 100, grps, 'Y1')

#> $`1`
#>            Y1       x_vars
#> 1   0.9541242 -4.075225250
#> 2   0.9509343 -3.998304986
#> 3   0.9475811 -3.921384722
#> 4   0.9440648 -3.844464458
#> 5   0.9403866 -3.767544194
#> 6   0.9365486 -3.690623931
#> 7   0.9325538 -3.613703667
#> 8   0.9284062 -3.536783403
#> 9   0.9241106 -3.459863139
#> 10  0.9196725 -3.382942876
#> 11  0.9150982 -3.306022612
#> 12  0.9103948 -3.229102348
#> 13  0.9055696 -3.152182084
#> 14  0.9006308 -3.075261820
#> 15  0.8955866 -2.998341557
#> 16  0.8904457 -2.921421293
#> 17  0.8852169 -2.844501029
#> 18  0.8799089 -2.767580765
#> 19  0.8745305 -2.690660501
#> 20  0.8690905 -2.613740238
#> 21  0.8635972 -2.536819974
#> 22  0.8580588 -2.459899710
#> 23  0.8524831 -2.382979446
#> 24  0.8468775 -2.306059183
#> 25  0.8412489 -2.229138919
#> 26  0.8356038 -2.152218655
#> 27  0.8299480 -2.075298391
#> 28  0.8242869 -1.998378127
#> 29  0.8186253 -1.921457864
#> 30  0.8129676 -1.844537600
#> 31  0.8073174 -1.767617336
#> 32  0.8016779 -1.690697072
#> 33  0.7960517 -1.613776808
#> 34  0.7904412 -1.536856545
#> 35  0.7848479 -1.459936281
#> 36  0.7792732 -1.383016017
#> 37  0.7737178 -1.306095753
#> 38  0.7681823 -1.229175490
#> 39  0.7626667 -1.152255226
#> 40  0.7571709 -1.075334962
#> 41  0.7516944 -0.998414698
#> 42  0.7462365 -0.921494434
#> 43  0.7407962 -0.844574171
#> 44  0.7353725 -0.767653907
#> 45  0.7299639 -0.690733643
#> 46  0.7245691 -0.613813379
#> 47  0.7191867 -0.536893115
#> 48  0.7138150 -0.459972852
#> 49  0.7084524 -0.383052588
#> 50  0.7030972 -0.306132324
#> 51  0.6977478 -0.229212060
#> 52  0.6924026 -0.152291797
#> 53  0.6870600 -0.075371533
#> 54  0.6817185  0.001548731
#> 55  0.6763765  0.078468995
#> 56  0.6710327  0.155389259
#> 57  0.6656858  0.232309522
#> 58  0.6603346  0.309229786
#> 59  0.6549780  0.386150050
#> 60  0.6496151  0.463070314
#> 61  0.6442450  0.539990578
#> 62  0.6388672  0.616910841
#> 63  0.6334809  0.693831105
#> 64  0.6280858  0.770751369
#> 65  0.6226817  0.847671633
#> 66  0.6172683  0.924591896
#> 67  0.6118457  1.001512160
#> 68  0.6064139  1.078432424
#> 69  0.6009733  1.155352688
#> 70  0.5955241  1.232272952
#> 71  0.5900669  1.309193215
#> 72  0.5846022  1.386113479
#> 73  0.5791308  1.463033743
#> 74  0.5736533  1.539954007
#> 75  0.5681706  1.616874271
#> 76  0.5626836  1.693794534
#> 77  0.5571932  1.770714798
#> 78  0.5517005  1.847635062
#> 79  0.5462065  1.924555326
#> 80  0.5407122  2.001475589
#> 81  0.5352186  2.078395853
#> 82  0.5297269  2.155316117
#> 83  0.5242381  2.232236381
#> 84  0.5187531  2.309156645
#> 85  0.5132730  2.386076908
#> 86  0.5077987  2.462997172
#> 87  0.5023310  2.539917436
#> 88  0.4968709  2.616837700
#> 89  0.4914190  2.693757964
#> 90  0.4859761  2.770678227
#> 91  0.4805427  2.847598491
#> 92  0.4751193  2.924518755
#> 93  0.4697064  3.001439019
#> 94  0.4643044  3.078359282
#> 95  0.4589134  3.155279546
#> 96  0.4535335  3.232199810
#> 97  0.4481649  3.309120074
#> 98  0.4428075  3.386040338
#> 99  0.4374611  3.462960601
#> 100 0.4321254  3.539880865
#> 
#> $`2`
#>            Y1       x_vars
#> 1   0.8031474 -4.075225250
#> 2   0.7972901 -3.998304986
#> 3   0.7914529 -3.921384722
#> 4   0.7856386 -3.844464458
#> 5   0.7798497 -3.767544194
#> 6   0.7740880 -3.690623931
#> 7   0.7683549 -3.613703667
#> 8   0.7626512 -3.536783403
#> 9   0.7569776 -3.459863139
#> 10  0.7513340 -3.382942876
#> 11  0.7457203 -3.306022612
#> 12  0.7401358 -3.229102348
#> 13  0.7345797 -3.152182084
#> 14  0.7290509 -3.075261820
#> 15  0.7235481 -2.998341557
#> 16  0.7180698 -2.921421293
#> 17  0.7126143 -2.844501029
#> 18  0.7071798 -2.767580765
#> 19  0.7017644 -2.690660501
#> 20  0.6963663 -2.613740238
#> 21  0.6909833 -2.536819974
#> 22  0.6856134 -2.459899710
#> 23  0.6802547 -2.382979446
#> 24  0.6749052 -2.306059183
#> 25  0.6695627 -2.229138919
#> 26  0.6642256 -2.152218655
#> 27  0.6588918 -2.075298391
#> 28  0.6535596 -1.998378127
#> 29  0.6482273 -1.921457864
#> 30  0.6428935 -1.844537600
#> 31  0.6375565 -1.767617336
#> 32  0.6322152 -1.690697072
#> 33  0.6268683 -1.613776808
#> 34  0.6215146 -1.536856545
#> 35  0.6161535 -1.459936281
#> 36  0.6107839 -1.383016017
#> 37  0.6054053 -1.306095753
#> 38  0.6000173 -1.229175490
#> 39  0.5946194 -1.152255226
#> 40  0.5892116 -1.075334962
#> 41  0.5837936 -0.998414698
#> 42  0.5783656 -0.921494434
#> 43  0.5729279 -0.844574171
#> 44  0.5674807 -0.767653907
#> 45  0.5620245 -0.690733643
#> 46  0.5565598 -0.613813379
#> 47  0.5510873 -0.536893115
#> 48  0.5456078 -0.459972852
#> 49  0.5401221 -0.383052588
#> 50  0.5346310 -0.306132324
#> 51  0.5291355 -0.229212060
#> 52  0.5236366 -0.152291797
#> 53  0.5181354 -0.075371533
#> 54  0.5126328  0.001548731
#> 55  0.5071299  0.078468995
#> 56  0.5016278  0.155389259
#> 57  0.4961274  0.232309522
#> 58  0.4906299  0.309229786
#> 59  0.4851361  0.386150050
#> 60  0.4796470  0.463070314
#> 61  0.4741634  0.539990578
#> 62  0.4686862  0.616910841
#> 63  0.4632160  0.693831105
#> 64  0.4577535  0.770751369
#> 65  0.4522993  0.847671633
#> 66  0.4468537  0.924591896
#> 67  0.4414172  1.001512160
#> 68  0.4359900  1.078432424
#> 69  0.4305723  1.155352688
#> 70  0.4251641  1.232272952
#> 71  0.4197654  1.309193215
#> 72  0.4143760  1.386113479
#> 73  0.4089955  1.463033743
#> 74  0.4036238  1.539954007
#> 75  0.3982601  1.616874271
#> 76  0.3929040  1.693794534
#> 77  0.3875548  1.770714798
#> 78  0.3822118  1.847635062
#> 79  0.3768739  1.924555326
#> 80  0.3715404  2.001475589
#> 81  0.3662101  2.078395853
#> 82  0.3608820  2.155316117
#> 83  0.3555549  2.232236381
#> 84  0.3502276  2.309156645
#> 85  0.3448989  2.386076908
#> 86  0.3395674  2.462997172
#> 87  0.3342319  2.539917436
#> 88  0.3288911  2.616837700
#> 89  0.3235435  2.693757964
#> 90  0.3181879  2.770678227
#> 91  0.3128229  2.847598491
#> 92  0.3074473  2.924518755
#> 93  0.3020598  3.001439019
#> 94  0.2966593  3.078359282
#> 95  0.2912447  3.155279546
#> 96  0.2858150  3.232199810
#> 97  0.2803694  3.309120074
#> 98  0.2749071  3.386040338
#> 99  0.2694274  3.462960601
#> 100 0.2639301  3.539880865
#> 
#> $`3`
#>            Y1       x_vars
#> 1   0.7692307 -4.075225250
#> 2   0.7634822 -3.998304986
#> 3   0.7577655 -3.921384722
#> 4   0.7520810 -3.844464458
#> 5   0.7464286 -3.767544194
#> 6   0.7408078 -3.690623931
#> 7   0.7352180 -3.613703667
#> 8   0.7296580 -3.536783403
#> 9   0.7241268 -3.459863139
#> 10  0.7186227 -3.382942876
#> 11  0.7131443 -3.306022612
#> 12  0.7076896 -3.229102348
#> 13  0.7022569 -3.152182084
#> 14  0.6968441 -3.075261820
#> 15  0.6914493 -2.998341557
#> 16  0.6860702 -2.921421293
#> 17  0.6807049 -2.844501029
#> 18  0.6753512 -2.767580765
#> 19  0.6700071 -2.690660501
#> 20  0.6646705 -2.613740238
#> 21  0.6593395 -2.536819974
#> 22  0.6540122 -2.459899710
#> 23  0.6486867 -2.382979446
#> 24  0.6433613 -2.306059183
#> 25  0.6380344 -2.229138919
#> 26  0.6327045 -2.152218655
#> 27  0.6273703 -2.075298391
#> 28  0.6220305 -1.998378127
#> 29  0.6166840 -1.921457864
#> 30  0.6113299 -1.844537600
#> 31  0.6059673 -1.767617336
#> 32  0.6005957 -1.690697072
#> 33  0.5952143 -1.613776808
#> 34  0.5898230 -1.536856545
#> 35  0.5844215 -1.459936281
#> 36  0.5790096 -1.383016017
#> 37  0.5735875 -1.306095753
#> 38  0.5681553 -1.229175490
#> 39  0.5627134 -1.152255226
#> 40  0.5572621 -1.075334962
#> 41  0.5518020 -0.998414698
#> 42  0.5463337 -0.921494434
#> 43  0.5408579 -0.844574171
#> 44  0.5353756 -0.767653907
#> 45  0.5298874 -0.690733643
#> 46  0.5243944 -0.613813379
#> 47  0.5188976 -0.536893115
#> 48  0.5133979 -0.459972852
#> 49  0.5078964 -0.383052588
#> 50  0.5023942 -0.306132324
#> 51  0.4968922 -0.229212060
#> 52  0.4913914 -0.152291797
#> 53  0.4858930 -0.075371533
#> 54  0.4803978  0.001548731
#> 55  0.4749068  0.078468995
#> 56  0.4694208  0.155389259
#> 57  0.4639407  0.232309522
#> 58  0.4584670  0.309229786
#> 59  0.4530005  0.386150050
#> 60  0.4475418  0.463070314
#> 61  0.4420912  0.539990578
#> 62  0.4366491  0.616910841
#> 63  0.4312158  0.693831105
#> 64  0.4257914  0.770751369
#> 65  0.4203759  0.847671633
#> 66  0.4149694  0.924591896
#> 67  0.4095715  1.001512160
#> 68  0.4041821  1.078432424
#> 69  0.3988008  1.155352688
#> 70  0.3934269  1.232272952
#> 71  0.3880600  1.309193215
#> 72  0.3826994  1.386113479
#> 73  0.3773442  1.463033743
#> 74  0.3719936  1.539954007
#> 75  0.3666466  1.616874271
#> 76  0.3613022  1.693794534
#> 77  0.3559594  1.770714798
#> 78  0.3506168  1.847635062
#> 79  0.3452734  1.924555326
#> 80  0.3399279  2.001475589
#> 81  0.3345790  2.078395853
#> 82  0.3292253  2.155316117
#> 83  0.3238656  2.232236381
#> 84  0.3184986  2.309156645
#> 85  0.3131230  2.386076908
#> 86  0.3077375  2.462997172
#> 87  0.3023408  2.539917436
#> 88  0.2969319  2.616837700
#> 89  0.2915096  2.693757964
#> 90  0.2860729  2.770678227
#> 91  0.2806210  2.847598491
#> 92  0.2751531  2.924518755
#> 93  0.2696685  3.001439019
#> 94  0.2641669  3.078359282
#> 95  0.2586480  3.155279546
#> 96  0.2531116  3.232199810
#> 97  0.2475581  3.309120074
#> 98  0.2419878  3.386040338
#> 99  0.2364015  3.462960601
#> 100 0.2308001  3.539880865
#> 
#> $`4`
#>            Y1       x_vars
#> 1   0.7416348 -4.075225250
#> 2   0.7360134 -3.998304986
#> 3   0.7304242 -3.921384722
#> 4   0.7248659 -3.844464458
#> 5   0.7193373 -3.767544194
#> 6   0.7138367 -3.690623931
#> 7   0.7083624 -3.613703667
#> 8   0.7029125 -3.536783403
#> 9   0.6974851 -3.459863139
#> 10  0.6920780 -3.382942876
#> 11  0.6866891 -3.306022612
#> 12  0.6813164 -3.229102348
#> 13  0.6759576 -3.152182084
#> 14  0.6706106 -3.075261820
#> 15  0.6652733 -2.998341557
#> 16  0.6599435 -2.921421293
#> 17  0.6546194 -2.844501029
#> 18  0.6492989 -2.767580765
#> 19  0.6439802 -2.690660501
#> 20  0.6386616 -2.613740238
#> 21  0.6333414 -2.536819974
#> 22  0.6280182 -2.459899710
#> 23  0.6226904 -2.382979446
#> 24  0.6173570 -2.306059183
#> 25  0.6120167 -2.229138919
#> 26  0.6066687 -2.152218655
#> 27  0.6013120 -2.075298391
#> 28  0.5959460 -1.998378127
#> 29  0.5905702 -1.921457864
#> 30  0.5851842 -1.844537600
#> 31  0.5797877 -1.767617336
#> 32  0.5743807 -1.690697072
#> 33  0.5689633 -1.613776808
#> 34  0.5635355 -1.536856545
#> 35  0.5580977 -1.459936281
#> 36  0.5526503 -1.383016017
#> 37  0.5471938 -1.306095753
#> 38  0.5417289 -1.229175490
#> 39  0.5362563 -1.152255226
#> 40  0.5307767 -1.075334962
#> 41  0.5252911 -0.998414698
#> 42  0.5198004 -0.921494434
#> 43  0.5143055 -0.844574171
#> 44  0.5088074 -0.767653907
#> 45  0.5033072 -0.690733643
#> 46  0.4978059 -0.613813379
#> 47  0.4923045 -0.536893115
#> 48  0.4868040 -0.459972852
#> 49  0.4813055 -0.383052588
#> 50  0.4758099 -0.306132324
#> 51  0.4703180 -0.229212060
#> 52  0.4648307 -0.152291797
#> 53  0.4593489 -0.075371533
#> 54  0.4538732  0.001548731
#> 55  0.4484042  0.078468995
#> 56  0.4429424  0.155389259
#> 57  0.4374884  0.232309522
#> 58  0.4320424  0.309229786
#> 59  0.4266046  0.386150050
#> 60  0.4211753  0.463070314
#> 61  0.4157544  0.539990578
#> 62  0.4103418  0.616910841
#> 63  0.4049373  0.693831105
#> 64  0.3995407  0.770751369
#> 65  0.3941515  0.847671633
#> 66  0.3887692  0.924591896
#> 67  0.3833932  1.001512160
#> 68  0.3780228  1.078432424
#> 69  0.3726571  1.155352688
#> 70  0.3672953  1.232272952
#> 71  0.3619364  1.309193215
#> 72  0.3565794  1.386113479
#> 73  0.3512231  1.463033743
#> 74  0.3458664  1.539954007
#> 75  0.3405081  1.616874271
#> 76  0.3351469  1.693794534
#> 77  0.3297816  1.770714798
#> 78  0.3244108  1.847635062
#> 79  0.3190334  1.924555326
#> 80  0.3136479  2.001475589
#> 81  0.3082531  2.078395853
#> 82  0.3028479  2.155316117
#> 83  0.2974310  2.232236381
#> 84  0.2920014  2.309156645
#> 85  0.2865581  2.386076908
#> 86  0.2811002  2.462997172
#> 87  0.2756268  2.539917436
#> 88  0.2701374  2.616837700
#> 89  0.2646315  2.693757964
#> 90  0.2591088  2.770678227
#> 91  0.2535692  2.847598491
#> 92  0.2480129  2.924518755
#> 93  0.2424403  3.001439019
#> 94  0.2368520  3.078359282
#> 95  0.2312490  3.155279546
#> 96  0.2256325  3.232199810
#> 97  0.2200041  3.309120074
#> 98  0.2143658  3.386040338
#> 99  0.2087198  3.462960601
#> 100 0.2030689  3.539880865
#> 
#> $`5`
#>            Y1       x_vars
#> 1   0.7134240 -4.075225250
#> 2   0.7079330 -3.998304986
#> 3   0.7024687 -3.921384722
#> 4   0.6970290 -3.844464458
#> 5   0.6916119 -3.767544194
#> 6   0.6862152 -3.690623931
#> 7   0.6808368 -3.613703667
#> 8   0.6754743 -3.536783403
#> 9   0.6701256 -3.459863139
#> 10  0.6647886 -3.382942876
#> 11  0.6594609 -3.306022612
#> 12  0.6541406 -3.229102348
#> 13  0.6488255 -3.152182084
#> 14  0.6435138 -3.075261820
#> 15  0.6382035 -2.998341557
#> 16  0.6328930 -2.921421293
#> 17  0.6275805 -2.844501029
#> 18  0.6222646 -2.767580765
#> 19  0.6169439 -2.690660501
#> 20  0.6116172 -2.613740238
#> 21  0.6062832 -2.536819974
#> 22  0.6009412 -2.459899710
#> 23  0.5955902 -2.382979446
#> 24  0.5902296 -2.306059183
#> 25  0.5848589 -2.229138919
#> 26  0.5794777 -2.152218655
#> 27  0.5740858 -2.075298391
#> 28  0.5686831 -1.998378127
#> 29  0.5632697 -1.921457864
#> 30  0.5578457 -1.844537600
#> 31  0.5524114 -1.767617336
#> 32  0.5469674 -1.690697072
#> 33  0.5415141 -1.613776808
#> 34  0.5360521 -1.536856545
#> 35  0.5305821 -1.459936281
#> 36  0.5251051 -1.383016017
#> 37  0.5196218 -1.306095753
#> 38  0.5141331 -1.229175490
#> 39  0.5086401 -1.152255226
#> 40  0.5031436 -1.075334962
#> 41  0.4976449 -0.998414698
#> 42  0.4921447 -0.921494434
#> 43  0.4866442 -0.844574171
#> 44  0.4811445 -0.767653907
#> 45  0.4756463 -0.690733643
#> 46  0.4701507 -0.613813379
#> 47  0.4646586 -0.536893115
#> 48  0.4591707 -0.459972852
#> 49  0.4536879 -0.383052588
#> 50  0.4482107 -0.306132324
#> 51  0.4427399 -0.229212060
#> 52  0.4372759 -0.152291797
#> 53  0.4318190 -0.075371533
#> 54  0.4263697  0.001548731
#> 55  0.4209281  0.078468995
#> 56  0.4154942  0.155389259
#> 57  0.4100682  0.232309522
#> 58  0.4046498  0.309229786
#> 59  0.3992388  0.386150050
#> 60  0.3938350  0.463070314
#> 61  0.3884377  0.539990578
#> 62  0.3830465  0.616910841
#> 63  0.3776608  0.693831105
#> 64  0.3722797  0.770751369
#> 65  0.3669025  0.847671633
#> 66  0.3615282  0.924591896
#> 67  0.3561559  1.001512160
#> 68  0.3507844  1.078432424
#> 69  0.3454127  1.155352688
#> 70  0.3400396  1.232272952
#> 71  0.3346638  1.309193215
#> 72  0.3292842  1.386113479
#> 73  0.3238994  1.463033743
#> 74  0.3185083  1.539954007
#> 75  0.3131095  1.616874271
#> 76  0.3077018  1.693794534
#> 77  0.3022840  1.770714798
#> 78  0.2968551  1.847635062
#> 79  0.2914138  1.924555326
#> 80  0.2859593  2.001475589
#> 81  0.2804905  2.078395853
#> 82  0.2750069  2.155316117
#> 83  0.2695077  2.232236381
#> 84  0.2639926  2.309156645
#> 85  0.2584611  2.386076908
#> 86  0.2529134  2.462997172
#> 87  0.2473494  2.539917436
#> 88  0.2417697  2.616837700
#> 89  0.2361750  2.693757964
#> 90  0.2305661  2.770678227
#> 91  0.2249444  2.847598491
#> 92  0.2193114  2.924518755
#> 93  0.2136692  3.001439019
#> 94  0.2080201  3.078359282
#> 95  0.2023666  3.155279546
#> 96  0.1967120  3.232199810
#> 97  0.1910598  3.309120074
#> 98  0.1854136  3.386040338
#> 99  0.1797780  3.462960601
#> 100 0.1741574  3.539880865
#> 
#> $`6`
#>             Y1       x_vars
#> 1   0.58556438 -4.075225250
#> 2   0.58026773 -3.998304986
#> 3   0.57496104 -3.921384722
#> 4   0.56964354 -3.844464458
#> 5   0.56431465 -3.767544194
#> 6   0.55897390 -3.690623931
#> 7   0.55362099 -3.613703667
#> 8   0.54825576 -3.536783403
#> 9   0.54287817 -3.459863139
#> 10  0.53748833 -3.382942876
#> 11  0.53208646 -3.306022612
#> 12  0.52667291 -3.229102348
#> 13  0.52124814 -3.152182084
#> 14  0.51581271 -3.075261820
#> 15  0.51036724 -2.998341557
#> 16  0.50491248 -2.921421293
#> 17  0.49944921 -2.844501029
#> 18  0.49397828 -2.767580765
#> 19  0.48850059 -2.690660501
#> 20  0.48301707 -2.613740238
#> 21  0.47752867 -2.536819974
#> 22  0.47203635 -2.459899710
#> 23  0.46654106 -2.382979446
#> 24  0.46104375 -2.306059183
#> 25  0.45554531 -2.229138919
#> 26  0.45004664 -2.152218655
#> 27  0.44454856 -2.075298391
#> 28  0.43905182 -1.998378127
#> 29  0.43355714 -1.921457864
#> 30  0.42806513 -1.844537600
#> 31  0.42257633 -1.767617336
#> 32  0.41709119 -1.690697072
#> 33  0.41161007 -1.613776808
#> 34  0.40613321 -1.536856545
#> 35  0.40066075 -1.459936281
#> 36  0.39519276 -1.383016017
#> 37  0.38972915 -1.306095753
#> 38  0.38426975 -1.229175490
#> 39  0.37881429 -1.152255226
#> 40  0.37336238 -1.075334962
#> 41  0.36791353 -0.998414698
#> 42  0.36246717 -0.921494434
#> 43  0.35702262 -0.844574171
#> 44  0.35157911 -0.767653907
#> 45  0.34613583 -0.690733643
#> 46  0.34069185 -0.613813379
#> 47  0.33524622 -0.536893115
#> 48  0.32979792 -0.459972852
#> 49  0.32434590 -0.383052588
#> 50  0.31888909 -0.306132324
#> 51  0.31342640 -0.229212060
#> 52  0.30795675 -0.152291797
#> 53  0.30247907 -0.075371533
#> 54  0.29699234  0.001548731
#> 55  0.29149557  0.078468995
#> 56  0.28598786  0.155389259
#> 57  0.28046839  0.232309522
#> 58  0.27493646  0.309229786
#> 59  0.26939148  0.386150050
#> 60  0.26383303  0.463070314
#> 61  0.25826084  0.539990578
#> 62  0.25267487  0.616910841
#> 63  0.24707527  0.693831105
#> 64  0.24146243  0.770751369
#> 65  0.23583701  0.847671633
#> 66  0.23019994  0.924591896
#> 67  0.22455248  1.001512160
#> 68  0.21889618  1.078432424
#> 69  0.21323296  1.155352688
#> 70  0.20756508  1.232272952
#> 71  0.20189519  1.309193215
#> 72  0.19622630  1.386113479
#> 73  0.19056183  1.463033743
#> 74  0.18490560  1.539954007
#> 75  0.17926180  1.616874271
#> 76  0.17363504  1.693794534
#> 77  0.16803030  1.770714798
#> 78  0.16245293  1.847635062
#> 79  0.15690861  1.924555326
#> 80  0.15140336  2.001475589
#> 81  0.14594346  2.078395853
#> 82  0.14053546  2.155316117
#> 83  0.13518610  2.232236381
#> 84  0.12990225  2.309156645
#> 85  0.12469089  2.386076908
#> 86  0.11955903  2.462997172
#> 87  0.11451363  2.539917436
#> 88  0.10956156  2.616837700
#> 89  0.10470953  2.693757964
#> 90  0.09996398  2.770678227
#> 91  0.09533106  2.847598491
#> 92  0.09081654  2.924518755
#> 93  0.08642577  3.001439019
#> 94  0.08216358  3.078359282
#> 95  0.07803426  3.155279546
#> 96  0.07404154  3.232199810
#> 97  0.07018849  3.309120074
#> 98  0.06647755  3.386040338
#> 99  0.06291050  3.462960601
#> 100 0.05948844  3.539880865
#>